fix(tv): ci round-trip fixes for sidebar overlay
- graphicsLayer imports from androidx.compose.ui.graphics (not ui.draw)
- restore TvSession/TvHome/TvSourcesScreen/TvPlayerScreen imports dropped
  in the TvApp rewrite
- inset px values passed as Float to BringIntoViewSpec factories
- drop duplicate Color import in DetailScreen, unused width import

+++ android/tv/src/main/java/com/crispy/tv/tv/ui/screens/HomeScreen.kt
@@ -145,7 +145,7 @@ fun HomeScreen(
 
         val rowsListState = rememberLazyListState()
         val density = LocalDensity.current
-        val rowsTopInsetPx = remember(density) { with(density) { RowsFocusTopInsetDp.roundToPx() } }
+        val rowsTopInsetPx = remember(density) { with(density) { RowsFocusTopInsetDp.roundToPx() }.toFloat() }
         val rowsScrollSpec = remember(rowsTopInsetPx) {
             tvRowsBringIntoViewSpec(
                 topInsetPx = rowsTopInsetPx,
@@ -334,7 +334,7 @@ private fun TvHomeRailRow(
 ) {
     val density = LocalDensity.current
     val isRtl = LocalLayoutDirection.current == LayoutDirection.Rtl
-    val startInsetPx = remember(density) { with(density) { TvHomeDimensions.EdgePadding.roundToPx() } }
+    val startInsetPx = remember(density) { with(density) { TvHomeDimensions.EdgePadding.roundToPx() }.toFloat() }
     val railScrollSpec = remember(startInsetPx, isRtl) {
         tvRailBringIntoViewSpec(startInsetPx = startInsetPx, isRtl = isRtl)
     }
